feat(asm): implement linear-scan register allocation with scanner and encoding functions
